+++ server/camcops_server/tasks/khandaker_mojo_medicationtherapy.py
@@ -26,6 +26,7 @@ camcops_server/tasks/khandaker_mojo_medicationtherapy.py
 
 """
 
+
 from typing import List, Optional, Type, TYPE_CHECKING
 
 from sqlalchemy.sql.sqltypes import Float, Integer, UnicodeText